requirements.txt
Appium-Python-Client
Faker
pytest
pytest-html
setuptools



